refactor: 역할에 따른 test 클래스 분리
